inference/requirements.txt
fal-client>=0.13.1
python-dotenv>=1.0.0
requests>=2.31.0
